Moved sorting to its own files
