+++ HousingSearchApi/V1/Gateways/SearchPersonESHelper.cs
@@ -4,6 +4,7 @@ using System.Threading.Tasks;
 using Amazon.Lambda.Core;
 using HousingSearchApi.V1.Domain;
 using HousingSearchApi.V1.Infrastructure;
+using HousingSearchApi.V1.Infrastructure.Sorting;
 using Nest;
 
 namespace HousingSearchApi.V1.Gateways
