+++ HousingSearchApi/V1/Infrastructure/Sorting/DefaultSort.cs
@@ -1,0 +1,12 @@
+using Nest;
+
+namespace HousingSearchApi.V1.Infrastructure.Sorting
+{
+    public class DefaultSort : IPersonListSort
+    {
+        public SortDescriptor<QueryablePerson> Get(SortDescriptor<QueryablePerson> descriptor)
+        {
+            return descriptor;
+        }
+    }
+}
