+++ HousingSearchApi/V1/Infrastructure/Sorting/IPersonListSort.cs
@@ -1,0 +1,9 @@
+using Nest;
+
+namespace HousingSearchApi.V1.Infrastructure.Sorting
+{
+    public interface IPersonListSort
+    {
+        SortDescriptor<QueryablePerson> Get(SortDescriptor<QueryablePerson> descriptor);
+    }
+}
