+++ HousingSearchApi/V1/Infrastructure/Sorting/IPersonListSortFactory.cs
@@ -1,0 +1,9 @@
+using HousingSearchApi.V1.Domain;
+
+namespace HousingSearchApi.V1.Infrastructure.Sorting
+{
+    public interface IPersonListSortFactory
+    {
+        IPersonListSort Create(GetPersonListRequest request);
+    }
+}
